module3: integration test of '/' endpoint
