+++ src/todolist/.env.test
@@ -5,6 +5,6 @@ FLASK_ENV=development
 # Change the following values for local development.
 TRELLO_KEY=pretendvalue
 TRELLO_TOKEN=fakestring
-BOARD_ID=5eecd4bae16d0019c3f56345
+BOARD_ID=5bbcd4bae16d0019c3f56345
 
 FLASK_SKIP_DOTENV=True
